rework hit to track association to not hammer the database quite so hard

+++ src/main/java/nocomment/master/db/TrackResume.java
@@ -5,9 +5,9 @@ import nocomment.master.util.ChunkPos;
 public class TrackResume {
     public final ChunkPos pos;
     public final short dimension;
-    public final long prevTrackID;
+    public final int prevTrackID;
 
-    TrackResume(int x, int z, short dimension, long prevTrackID) {
+    TrackResume(int x, int z, short dimension, int prevTrackID) {
         this.pos = new ChunkPos(x, z);
         this.dimension = dimension;
         this.prevTrackID = prevTrackID;
